Move issue detection to GitService
